requirements.txt
Flask
Flask-CORS
openai
python-dotenv
requests
gunicorn
